Added mysql-based deduping of received SHA1
This is an attempt to remove duplicate emails.  These emails happen when (for
instance) a merge occurs, and somehow, github pushes parts of both parents at
us even though we already saw one of them before.  This has been annoying a
number of us for a while, and this is my first shot at curing the problem.

Also split the config out into a separate file